refactor: multitask model decorator handling of batch shape checks
Previously, the LMC decorator blindly ate any ValueErrors from the
superclass `_check_batch_shape` method. Now, the specific check that
was trying to be caught is its own method, and the LMC decorator simply
overrides this method so that the error isn't raised in the first place.

Also refactored the tests for this - instead of two distinct tests,
there's now just one parametrised test.

Refs: #358