feat: convert integration tests to pytest; run with ndarray and tensor
We run with both NDArray and Tensor inputs as a rough test that the
client interface is indifferent to whether inputs are NDArrays or
Tensors.

I haven't touched test_multitask_usage.py as its only test is skipped.

This does significantly increase the runtime of the integration tests;
not only are we running everything twice, but we are now checking all
combinations of partitioner and aggregator in test_distribute_usage.py.
This latter point can be reverted if needed.